refactor: useData 훅을 ApiContext.tsx로 통합하고 불필요한 파일 삭제
